`parseMail`: differentiate between empty and missing body
The returned `body.html` and `body.text` are now `null` if no corresponding
body part is found.
The empty string is returned only if the body is present but empty.

Previously, the empty string was returned also for messages without
body parts, and it was not possible to differentiate the two cases.